Style fixes

+++ src/Orbit/ShaderGen/Variables/Variable.h
@@ -41,8 +41,9 @@ namespace ShaderGen
 
 	public:
 
-		/* Stores the value in a local variable. Useful when we want to manipulate swizzles within
-		 * a variable, since `Vec2(1.0, 0.5).g *= 2.0;` is ill-behaved. */
+		/** Stores the value in a local variable. Useful when we want to manipulate swizzles within a variable, since
+		 * `Vec2(1.0, 0.5).g *= 2.0;` is ill-behaved.
+		 */
 		void StoreValue( void );
 
 		/** Fetch the value of this variable. Marks this variable as used. */
@@ -62,8 +63,8 @@ namespace ShaderGen
 		Variable operator/ ( const Variable& rhs )   const;
 		Variable operator+ ( const Variable& rhs )   const;
 		Variable operator- ( const Variable& rhs )   const;
-		Variable operator- ( void )                   const;
-		Variable operator[]( size_t index )           const;
+		Variable operator- ( void )                  const;
+		Variable operator[]( int index )             const;
 
 		SwizzlePermutations* operator->( void ) const;
